cloud_functions/requirements.txt
requests
google-cloud-storage
functions-framework
